backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.3
python-dotenv==1.0.0
pydantic[email]==2.7.4
python-multipart==0.0.12
ultralytics==8.2.100
opencv-python==4.10.0.84
numpy==1.26.4
